chore: refactor test suites
