+++ apps/backend-relayer/package.json
@@ -17,7 +17,8 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "test:e2e": "jest --config ./test/e2e/jest-e2e.json",
+    "test:integrations": "jest --config ./test/integrations/jest-e2e.json",
     "prisma:migrate": "prisma migrate deploy"
   },
   "dependencies": {
