+++ apps/backend-relayer/test/e2e/admin.e2e-spec.ts
@@ -1,6 +1,6 @@
 import request from 'supertest';
 import { INestApplication } from '@nestjs/common';
-import { createTestingApp } from './setups/create-app';
+import { createTestingApp } from '../setups/create-app';
 
 describe('Admin E2E', () => {
   let app: INestApplication;
