+++ apps/backend-relayer/test/e2e/jest-e2e.json
@@ -1,6 +1,6 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": ".",
+  "rootDir": "../",
   "testEnvironment": "node",
   "testRegex": ".e2e-spec.ts$",
   "transform": {
@@ -8,7 +8,7 @@
   },
   "setupFilesAfterEnv": [],
   "globalSetup": "<rootDir>/setups/jest-e2e.setup.ts",
-  "globalTeardown": "<rootDir>/setups/jest-e2e.teardown.ts",
+  "globalTeardown": "<rootDir>/setups/jest.teardown.ts",
   "moduleNameMapper": {
     "^@prisma/(?!client)(.*)$": "<rootDir>/../prisma/$1",
     "^@libs/(.*)$": "<rootDir>/../src/libs/$1"
