+++ apps/backend-relayer/test/integrations/jest-e2e.json
@@ -1,0 +1,16 @@
+{
+  "moduleFileExtensions": ["js", "json", "ts"],
+  "rootDir": "../",
+  "testEnvironment": "node",
+  "testRegex": ".e2e-integration.ts$",
+  "transform": {
+    "^.+\\.(t|j)s$": "ts-jest"
+  },
+  "setupFilesAfterEnv": [],
+  "globalSetup": "<rootDir>/setups/jest-integrations.setup.ts",
+  "globalTeardown": "<rootDir>/setups/jest.teardown.ts",
+  "moduleNameMapper": {
+    "^@prisma/(?!client)(.*)$": "<rootDir>/../prisma/$1",
+    "^@libs/(.*)$": "<rootDir>/../src/libs/$1"
+  }
+}
