+++ apps/backend-relayer/test/setups/seed.ts
@@ -65,3 +65,20 @@ export const seedDB = async (
     await prisma.$disconnect();
   }
 };
+
+export const seedDBe2e = async () => {
+  const prisma = new PrismaClient();
+  try {
+    await prisma.$connect();
+
+    await seedAdmin(prisma, 'admin@x.com', 'ChangeMe123!');
+
+    await prisma.$disconnect();
+
+    console.log('Seeding completed.');
+  } catch (error) {
+    console.error('Error seeding db:', error);
+  } finally {
+    await prisma.$disconnect();
+  }
+};
